Add files via upload
a. Overcrowding of the places is avoided by allowing only a limited number of people at a time.
b. Users can pre check whether the shop is open or not at the specific time and can pre check
whether the required product is available or not.
c. Access to the location will help users to get a précised list of shops.
d. As a concern of security, all the details of customers will be stored in the database.
e. The booking through calling mechanism in the system will overcome the system down problems.
f. Service providers will get the count of the exact number of people present at the shop at a time.
g. Administrator can monitor the database and can have control over it.
h. Cost efficient for users as no charges will be charged for booking purposes.
i. For virtual queue management, a system is designed where the customers can pre-book their time
slots for visiting the malls, restaurants or any kind of shops. 
The system will help customers to avoid the inconvenience like overcrowded places.